reportes listos

+++ Models/SalesReportRowModel.cs
@@ -6,9 +6,9 @@ namespace RestaurantJapanese.Models
     {
         public int IdTicket { get; set; }
         public DateTime CreatedAt { get; set; }
-        public int Subtotal { get; set; }
-        public int Tax { get; set; }
-        public int Total { get; set; }
+        public decimal Subtotal { get; set; } // Cambiado de int a decimal
+        public decimal Tax { get; set; }      // Cambiado de int a decimal
+        public decimal Total { get; set; }    // Cambiado de int a decimal
 
         // Propiedades calculadas para la UI
         public string FormattedDate => CreatedAt.ToString("dd/MM/yyyy HH:mm");
